requirements.txt
certifi==2020.4.5.1
discord.py==1.3.3
prismapy==2.1.2
psutil==5.7.0
requests==2.23.0
praw==7.1.0
